requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.6b0
CacheControl==0.12.6
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.2
click==8.0.1
colorama==0.4.4
contextlib2==21.6.0
cycler==0.10.0
distlib==0.3.2
distro==1.5.0
dj-database-url==0.5.0
Django==3.2.5
django-ckeditor==6.1.0
django-extensions==3.1.3
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.4
flake8==3.9.2
gunicorn==20.1.0
html5lib==1.1
idna==3.2
ipaddr==2.2.0
kiwisolver==1.3.1
lockfile==0.12.2
matplotlib==3.4.2
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
numpy==1.21.0
packaging==21.0
pathspec==0.8.1
pep517==0.10.0
Pillow==8.3.1
progress==1.5
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
retrying==1.3.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
whitenoise==5.2.0
